sequencer pattern can now be resized

+++ AAudioTrack2/src/main/java/smallville7123/aaudiotrack2/PatternGroup.java
@@ -13,17 +13,17 @@ public class PatternGroup<P extends PatternList<? extends Pattern>> {
 
     public P newPatternList(P patternList) {
         patternList.DAWReference = DAWReference;
-        patternList.nativePatternList = DAWReference.createPatternList();
+        if (DAWReference != null) {
+            patternList.nativePatternList = DAWReference.createPatternList();
+        }
         patternListArrayList.add(patternList);
         return patternList;
     }
 
     public void removePatternList(P patternList) {
-        DAWReference.deletePatternList(patternList.nativePatternList);
+        if (DAWReference != null) {
+            DAWReference.deletePatternList(patternList.nativePatternList);
+        }
         patternListArrayList.remove(patternList);
-    }
-
-    public void setDAW(AAudioTrack2 audioTrack) {
-        DAWReference = audioTrack;
     }
 }
